Checkpoint: typegen correct generation of project file
 - TODO: correct detection of CapoBundle subclasses, for accurate
   instantiation of capo-dependent modules (when they import
   modules provided by the Capo bundle)
 - doesn't write a new project file if the contents are identical
 - writes a project backup file only once at most, only if the
   project file is being updated (TODO: improve for iterating in watch mode)
 - watches project file for changes (trigger another build, if in watch mode)
 - separates "import ts from js" for `vitest` to its own resolveId hook
 - moves .hlbundle registration to a load hook
 - detect project-file importName from .hlbundle content during first load
   requires 'export default class ...' in .hlbundle files
 - use program-name for importName once the bundle has reached the
   geneneration where it is compiled into the project
 -

+++ src/helios/StellarHeliosProject.ts
@@ -25,6 +25,7 @@ type HeliosBundleTypes = {
 
 type bundleWithStatus = {
     filename: string;
+    importName: string;
     bundle?: Option<HeliosScriptBundle>;
     types?: Option<HeliosBundleTypes>;
 };
@@ -43,20 +44,18 @@ type bundleWithStatus = {
  * do the above.
  */
 export class StellarHeliosProject {
-    static root: string
-    static loadExistingProject() : Promise<StellarHeliosProject> | undefined {
-
+    static root: string;
+    static loadExistingProject(): Promise<StellarHeliosProject> | undefined {
         const root = StellarHeliosProject.findProjectRoot();
         if (existsSync(`${root}/hlproject.mjs`)) {
-            return  import(`${root}/hlproject.mjs`).then((projectPackage) => {
+            return import(`${root}/hlproject.mjs`).then((projectPackage) => {
                 if (!projectPackage.project) {
                     throw new Error(`hlproject.mjs must export a \`project\``);
                 }
-                return  projectPackage.project;
-            })
+                return projectPackage.project;
+            });
         }
     }
-
     bundles: Map<string, bundleWithStatus>;
     capoBundle: HeliosScriptBundle | null = null;
     projectRoot: string;
@@ -64,16 +63,35 @@ export class StellarHeliosProject {
         this.bundles = new Map();
         this.projectRoot = StellarHeliosProject.findProjectRoot();
     }
-
+    addBundleWithMockTypes(filename: string) {
+        this.bundles.set(filename, { filename, importName: this.getImportNameFromHlBundle(filename) });
+        // probably the generic types of the mkDatum, mkRedeemer, and readDatum proxies
+        // in the base class will be enough to bootstrap the real types, in which those proxies
+        // get a more specific type defined (via the generated .d.ts, from the underlying helios source) 
+        // for each bundle file.
+        const mockTypesAreImportant = false;
+        if (mockTypesAreImportant) {
+            this.writeMockTypes(filename);
+        }
+    }
     // call from code-generated hlproject.mjs with instantiated bundle
     // call from rollup plugin with bundle filename
-    addBundle(filename: string, bundleClass?: typeof HeliosScriptBundle) {
+    addBundle(
+        absoluteFilename: string,
+        bundleClass?: typeof HeliosScriptBundle
+    ) {
+        // if the file location is within the project root, make it relative
+        // otherwise, use the absolute path
+        const filename = absoluteFilename.startsWith(this.projectRoot)
+            ? path.relative(this.projectRoot, absoluteFilename)
+            : absoluteFilename;
+
         if (bundleClass) {
             let bundle: HeliosScriptBundle | undefined;
             // if the bundle has a CapoBundle, use it
-            const isCapoBundle = false
-            debugger
-            if (isCapoBundle) { //bundleClass.prototype instanceof CapoBundle) {
+            const isCapoBundle = false;
+            if (isCapoBundle) {
+                //bundleClass.prototype instanceof CapoBundle) {
                 if (this.bundles.size > 0) {
                     if (this.capoBundle) {
                         throw new Error(
@@ -81,7 +99,7 @@ export class StellarHeliosProject {
                         );
                     } else {
                         throw new Error(
-                            `the CapoBundle must be registered before any other bundles (at ${filename}`
+                            `the CapoBundle must be registered before any other bundles (at ${absoluteFilename}`
                         );
                     }
                 }
@@ -89,16 +107,43 @@ export class StellarHeliosProject {
             } else {
                 if (!this.capoBundle) {
                     throw new Error(
-                        `the CapoBundle must be registered before any other bundles (at ${filename}`
+                        `the CapoBundle must be registered before any other bundles (at ${absoluteFilename}`
                     );
                 }
                 bundle = new (bundleClass as any)(this.capoBundle);
             }
             const types = bundle ? this.genTypes(bundle) : undefined;
-            this.bundles.set(filename, { filename, bundle, types });
+            const importName =
+                bundle?.program.name ||
+                this.getImportNameFromHlBundle(filename);
+
+            this.bundles.set(filename, { filename, bundle, types, importName });
         } else {
-            this.bundles.set(filename, { filename });
+            const importName = this.getImportNameFromHlBundle(filename);
+
+            this.bundles.set(filename, { filename, importName });
         }
+        this.writeProjectFile();
+    }
+
+    private getImportNameFromHlBundle(filename: string) {
+        const fileContent = readFileSync(filename, "utf-8");
+        const importNameMatch = fileContent.match(
+            /export\s+default\s+(?:class|function)\s+([a-zA-Z0-9_]+)/
+        );
+        if (!importNameMatch) {
+            throw new Error(
+                `could not extract **default export** name from ${filename}\n` +
+                    `  expected: export default class ...`
+            );
+        }
+        const importName = importNameMatch[1];
+        if (!importName.match(/^[a-zA-Z0-9_]+$/)) {
+            throw new Error(
+                `invalid import name from ${filename}: ${importName}`
+            );
+        }
+        return importName;
     }
 
     hasBundleClass(filename: string) {
@@ -107,21 +152,33 @@ export class StellarHeliosProject {
         }
     }
 
+    writeMockTypes(filename: string) {
+        // const typeFilename = filename.replace(/\.hlbundle\.js$/, ".hlbundle.d.ts");
+        // writeFileSync(typeFilename, `// generated by StellarHeliosProject using Stellar Helios Rollup type-generator
+        //     // mock types
+    }
+
     writeTypeInfo(filename: string) {
         const bundle = this.bundles.get(filename);
         if (!bundle) {
             throw new Error(`bundle not found: ${filename}`);
-        } else if( !bundle.bundle ) {
-            throw new Error(`cannot write type info for ${filename} for newly-added bundle (check for hasBundleClass first)`);
+        } else if (!bundle.bundle) {
+            throw new Error(
+                `cannot write type info for ${filename} for newly-added bundle (check for hasBundleClass first)`
+            );
         }
         const types = this.genTypes(bundle.bundle);
-        const typeFilename = filename.replace(/\.hlbundle\.js$/, ".hlbundle.d.ts");
+        const typeFilename = filename.replace(
+            /\.hlbundle\.js$/,
+            ".hlbundle.d.ts"
+        );
         const className = bundle.bundle.constructor.name;
 
-        const parentClass = "Placehodler"// bundle.bundle?.constructor.name || "HeliosScriptBundle";
-        debugger
+        const parentClass = "Placehodler"; // bundle.bundle?.constructor.name || "HeliosScriptBundle";
         throw new Error(`todo: write type info to ${typeFilename}`);
-        writeFileSync(typeFilename, `// generated by StellarHeliosProject using Stellar Helios Rollup type-generator
+        writeFileSync(
+            typeFilename,
+            `// generated by StellarHeliosProject using Stellar Helios Rollup type-generator
 import type {CapoBundle} from "src/CapoHeliosBundle.ts"   // todo import  from @stellar-contracts
 import type {HeliosScriptBundle} from "src/helios/HeliosScriptBundle.ts" // todo import from @stellar-contracts
 
@@ -136,7 +193,8 @@ export default class ${className} extends ${parentClass} {
         placeholder: "show proxy types here";
     }
 }
-           `);
+           `
+        );
     }
 
     genTypes(bundle: HeliosScriptBundle): HeliosBundleTypes {
@@ -147,38 +205,57 @@ export default class ${className} extends ${parentClass} {
         // const [canonicalType, permissiveType] = genTypes(bundle)
     }
 
-    generateProject() {
+    private _didCreateBackupFile = false;
+    createBackupFileOnce() {
+        // uses this.projectFilename
+        // creates a backup of the existing hlproject.mjs
+        // if this object already did this, it isn't done a second time
+
+        if (!this._didCreateBackupFile) {
+            if (existsSync(this.projectFilename)) {
+                writeFileSync(
+                    `${this.projectFilename}.bk`,
+                    readFileSync(this.projectFilename)
+                );
+            }
+            this._didCreateBackupFile = true;
+        }
+    }
+
+    writeProjectFile() {
         // creates or updates hlproject.mjs in the project root
         // makes a backup of the existing hlproject.mjs
         // writes the new hlproject.mjs
 
-        if (existsSync(`${this.projectRoot}/hlproject.mjs`)) {
-            writeFileSync(
-                `${this.projectRoot}/hlproject.mjs.bk`,
-                readFileSync(`${this.projectRoot}/hlproject.mjs`)
-            );
-        }
+        this.createBackupFileOnce();
+
         let content = `// generated by StellarHeliosProject using Stellar Helios Rollup type-generator
 
-import { HeliosScriptBundle } from "src/helios/HeliosScriptBundle.ts"
-import { CapoHeliosBundle } from "src/CapoHeliosBundle.ts"
-import { StellarHeliosProject } from "src/helios/StellarHeliosProject.ts"
+import { HeliosScriptBundle } from "src/helios/HeliosScriptBundle.ts"   // todo import from @stellar-contracts
+import { CapoHeliosBundle } from "src/CapoHeliosBundle.ts"                    // todo import from @stellar-contracts
+import { StellarHeliosProject } from "src/helios/StellarHeliosProject.ts" // todo import from @stellar-contracts
 `;
         for (const [filename, bundle] of this.bundles) {
-            content += `import ${
-                bundle.bundle?.program.name || "unk"
-            } from "${filename}";\n`;
+            content += `import ${bundle.importName} from "${filename}";\n`;
         }
         content += `\nexport const project = new StellarHeliosProject();\n\n`;
         for (const [filename, bundle] of this.bundles) {
-            content += `project.addBundle("${filename}", ${
-                bundle.bundle?.program.name || "unk"
-            });\n`;
+            content += `project.addBundle("${filename}", ${bundle.importName});\n`;
         }
-        writeFileSync(`${this.projectRoot}/hlproject.mjs`, content);
-        console.log(
-            `📦 StellarHeliosProject: wrote hlproject.mjs to ${this.projectRoot}`
-        );
+        // compare the content to the existing file
+        // ONLY if it is different, write the new file
+        let existingContent = "--- not existing ---";
+        if (existsSync(this.projectFilename)) {
+            existingContent = readFileSync(this.projectFilename, "utf-8");
+        }
+        if (existingContent === content) {
+            console.log(`📦 StellarHeliosProject: no changes to hlproject.mjs`);
+        } else {
+            writeFileSync(`${this.projectRoot}/hlproject.mjs`, content);
+            console.log(
+                `📦 StellarHeliosProject: wrote hlproject.mjs to ${this.projectRoot}`
+            );
+        }
     }
 
     static findProjectRoot() {
@@ -208,7 +285,14 @@ import { StellarHeliosProject } from "src/helios/StellarHeliosProject.ts"
             }
         }
         console.log(`📦 StellarHeliosProject: found project root at ${dir}`);
-        this.root = dir
+        this.root = dir;
         return dir;
+    }
+
+    static get projectFilename() {
+        return this.findProjectRoot() + "/hlproject.mjs";
+    }
+    get projectFilename() {
+        return StellarHeliosProject.projectFilename;
     }
 }
